Move habilitations routes tests to specific file
